requirements.txt
streamlit>=1.28.0
supabase>=2.0.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
pandas>=2.0.0
plotly>=5.15.0
openpyxl>=3.1.0